EHTNCDBase

Documentation for EHTNCDBase.

EHTNCDBase.ncdmodes

EHTNCDBase.define_group

EHTNCDBase.get_group

EHTNCDBase.split_group

EHTNCDBase.ncdmodes — Constant

ncdmodes::NamedTuple

A named tuple relating Symbols to actual strings for the access mode to netCDF files using NCDatasets.jl. Keys are:

Keys and Values

:read: open an existing netCDF file or OPeNDAP URL in read-only mode ("r" in NCDatasets.jl).

:create: create a new NetCDF file (an existing file with the same name will be overwritten; "c" in NCDatasets.jl)

:append: open filename into append mode (i.e. existing data in the netCDF file is not overwritten; "a" in NCDatasets.jl)

EHTNCDBase.define_group — Method

define_group(dataset, groups) -> NCDataset

Define the given group in the given dataset.

Arguments

dataset::NCDataset: Input dataset

groups::AbstractVector: A vector of strings for the path of the group. Assuming that this is an output of split_group.

EHTNCDBase.get_group — Method

get_group(dataset, groups) -> NCDataset

Get the given group specified in the given NCDataset.

Arguments

dataset::NCDataset: Input dataset

groups::AbstractVector: A vector of strings for the path of the group. Assuming that this is an output of split_group.

EHTNCDBase.split_group — Method

split_group(group) -> Vector{String}

Purse the input path of the group and split it into a vector.

Arguments

group::AbstractString: Input path of the group

Examples

julia> split_group("rootgroup/subgroup/subsubgroup")
3-element Vector{String}:
 "rootgroup"
 "subgroup"
 "subsubgroup"